refactor: extract schema constants and fix cross-version comment detection
- Extract SQL schema constants, migration helpers, and DDL strings from
  archive/index.py into a new archive/_schema.py module; index.py now
  re-exports them for backward compatibility
- Fix _is_daydream_comment to match against a version-stable footer
  prefix (_DAYDREAM_FOOTER_PREFIX) instead of the full version-pinned
  DAYDREAM_FOOTER, so comments posted by any daydream release are
  correctly identified
- Fix pr_link_signal to extract pr_repo from the PR payload's
  head.repo.full_name rather than always falling back to repo_slug,
  and guard against missing pr_number with a continue
- Add test covering dry_run=True orphan re-link: verifies in-memory
  linkage preview fires build_annotation but suppresses the DB write

Daydream-Run: 20260606184742-495a7163
Daydream-Version: 0.19.0
